move R2 implementation to main package
