+++ examples/r2-image-server/README.md
@@ -1,28 +1,11 @@
-# simple-json-server
+# r2-image-server
 
-* A simple HTTP JSON server implemented in Go and compiled with tinygo.
+* An example server which returns image from Cloudflare R2.
+* This server is implemented in Go and compiled with tinygo.
 
 ## Example
 
-* https://simple-json-server.syumai.workers.dev
-
-### Request
-
-```
-curl --location --request POST 'https://simple-json-server.syumai.workers.dev/hello' \
---header 'Content-Type: application/json' \
---data-raw '{
-    "name": "syumai"
-}'
-```
-
-### Response
-
-```json
-{
-    "message": "Hello, syumai!"
-}
-```
+* https://r2-image-server.syumai.workers.dev/syumai.png
 
 ## Development
 
@@ -32,8 +15,6 @@ This project requires these tools to be installed globally.
 
 * wrangler
 * tinygo
-* [easyjson](https://github.com/mailru/easyjson)
-  - `go install github.com/mailru/easyjson/...@latest`
 
 ### Commands
