+++ examples/r2-image-server/main.go
@@ -22,7 +22,7 @@ func handleErr(w http.ResponseWriter, msg string, err error) {
 // This example is based on implementation in syumai/workers-playground
 // * https://github.com/syumai/workers-playground/blob/e32881648ccc055e3690a0d9c750a834261c333e/r2-image-viewer/src/index.ts#L30
 func handler(w http.ResponseWriter, req *http.Request) {
-	bucket, err := NewR2Bucket(bucketName)
+	bucket, err := workers.NewR2Bucket(bucketName)
 	if err != nil {
 		handleErr(w, "failed to get R2Bucket\n", err)
 		return
