+++ r2object.go
@@ -1,4 +1,4 @@
-package main
+package workers
 
 import (
 	"errors"
